pyproject.toml
[project]
name = "azure-ml-lora-finetune"
version = "0.1.0"
description = "LoRA fine-tuning for small language models on Azure ML"
requires-python = ">=3.12"
dependencies = [
    "torch==2.9.1",
    "transformers==4.57.3",
    "peft==0.18.0",
    "bitsandbytes==0.45.0",
    "accelerate==1.12.0",
    "datasets==4.4.1",
    "pyyaml==6.0.3",
    "python-dotenv==1.2.1",
    "azure-storage-blob==12.19.0",
    "azure-ai-ml==1.30.0",
    "azure-identity==1.25.1",
]
